==> 09 llm annotation quality analysis/01_create_subsets_for_annotation.ipynb <==
## Notebook: Create Subset for Annotation of Synthetic Examples

## Packages

import pandas as pd
import random
import json

## Settings

N_FOLDS = 6
MODEL = "Llama3_70B"
FEW_SHOT = "fixed"
N_SAMPLES_FROM_SPLIT = 100
SEED = 43

random.seed(SEED)

## Code

annotation_examples = []
for fold_id in range(N_FOLDS):
    with open(f"../07 train models/synth/{MODEL}/{FEW_SHOT}/split_{fold_id}.json", 'r') as json_file:
        synthetic_data = random.sample(json.load(json_file), N_SAMPLES_FROM_SPLIT)
        for example in synthetic_data:
            example["few_shot_condtion"] = FEW_SHOT
            annotation_examples.append(example)

df = pd.DataFrame(annotation_examples, columns=['id', 'text', 'model', 'split', 'few_shot_condtion'])
df.to_csv(f"annotation_datasets/synth_annotation_labelstudio_input/annotation_{MODEL}_{FEW_SHOT}_{N_SAMPLES_FROM_SPLIT*N_FOLDS}.csv")
