==> 09 llm annotation quality analysis/02_merge_examples_for_annotation.ipynb <==
# Notebook: Create Confirmation Subsets

## Packages

import pandas as pd
import json
import os

## Settings

N_PARTS = 1

OUTPUT_PATH_PREFIX = "annotation_datasets/annotation_jakob"

## Code

input_folder = 'annotation_datasets/synth_annotation_nils_labelstudio_output'

all_files = [file for file in os.listdir(input_folder) if file.endswith('.csv')]
df = pd.concat([pd.read_csv(os.path.join(input_folder, file)) for file in all_files], ignore_index=True)

df

df["second_annotator_comment"] = None
df["two_or_more_sentences"].unique()

def merge_labels_implicit_to_explicit(dataframe):
    # Benutzerdefinierte Funktion, um label-implicit zu label-explicit hinzuzufügen und 'type' hinzufügen
    def merge_labels(row):
        labels_explicit = row["label-explicit"]
        labels_implicit = row["label-implicit"]

        if isinstance(labels_explicit, str) == False:
            labels_explicit = []
        else:
            labels_explicit = json.loads(labels_explicit)

        if isinstance(labels_implicit, str) == False:
            labels_implicit = []
        else:
            labels_implicit = json.loads(labels_implicit)

        labels_total = ""
        for annotation in labels_implicit:
            tuple = (annotation["labels"][0][:-len("-no-phrase")].rsplit('-', 1)[0],
                     annotation["labels"][0][:-len("-no-phrase")].rsplit('-', 1)[-1], 'NULL')
            labels_total += str(tuple) + "\n"

        for annotation in labels_explicit:
            tuple = (annotation["labels"][0].rsplit('-', 1)[0],
                     annotation["labels"][0].rsplit('-', 1)[-1], annotation["text"])
            labels_total += str(tuple) + "\n"

        row["annotation"] = labels_total

        # Sonderfälle
        # Update: Text is a bit misleading - my fault... However, I wanted to mark examples with mentioned aspects towards which no sentiment
        # was expressed by the author.
        if row["aspect_available_without_judgement"] == 'An aspect is addressed in the text without an explicit mention by the author':
            row["aspect_available_without_judgement"] = "✅"
        else:
            row["aspect_available_without_judgement"] = ""

        if row["two_or_more_sentences"] == "Two or more sentences":
            row["annotation"] = "⚠️ Mehr als 1 Satz im Text identifiziert von Annoator A"

        row["second_annotator_comment"] = ""
        return row

    dataframe = dataframe.apply(merge_labels, axis=1)

    return dataframe

total_rows = len(df)
rows_per_part = total_rows // N_PARTS

for idx in range(N_PARTS):
    start_idx = idx * rows_per_part
    end_idx = (idx + 1) * rows_per_part if idx < N_PARTS - 1 else total_rows
    part_df = df.iloc[start_idx:end_idx]

    filename = f"{OUTPUT_PATH_PREFIX}_{idx}.csv"

    part_df = merge_labels_implicit_to_explicit(part_df)
    part_df = part_df[["id", "second_annotator_comment", "annotation", "aspect_available_without_judgement", "text", "model", "few_shot_condtion", "split"]]

    part_df.to_csv(filename, index=False)

    print(f"Part {idx} saved to '{filename}'.")

result_df = part_df[part_df.duplicated(subset=['text'], keep=False) & part_df.groupby('text')['annotation'].transform('nunique').gt(1)]
result_df[0:20]

for (text_value, model_value, few_shot_condition_value), group_df in result_df.groupby(['text', 'model', 'few_shot_condtion']):
    unique_annotation_values = group_df['annotation'].unique()
    print(f"Text: {text_value}, Model: {model_value}, Few Shot Condition: {few_shot_condition_value}, Possible Annotation Values: {', '.join(unique_annotation_values)}")


# Todo: Restaurant als aspekt begriff bei ambiente
